wip, more specific exceptions for ssh failures
The current AnsibleConnectionFailure can be raised in
multiple places, and sometimes with the same msg.
This makes it harder to track down the cause of an error.

This adds a sub class of AnsibleConnectionFailure for the
ssh connection plugin, and then two further subclasses for
the two scenarios that can triggered it.

Next step is to update the msg's used to make them more specific.
Then add any useful context info available to the exception
to it's data (for ex, the error code, or the host connecting to,
etc)