Use stdlib logging
hookup logging hopefully everywhere

Start flushing out custom logging setup

Support ANSIBLE_DEBUG env or config for stderr logging.

Add a custom handler that falls back to NullHandler and
sets up our logging filters.

Add logging filter for ElevateExceptionToCriticalLoggingFilter,
that will increase the log level of log.exception(e) use from
ERROR to CRITICAL. That's likely overkill, but it is an interesting
example of a filter.

Add example of quieting specific loggers without effect default
log level.

Add lots of notes and todos.

Add a AnsibleLogger with our filters

Use logging.setLoggerClass() to make it the default logger type

add HostLoggerAdapter and use it in connection plugin

add colorized console debug logger

Add levels module with log level data
import V style levels from logger.levels
Add debug logger that emulates display.debug

Add DefaultAttributesFilter

This is a logging filter that will populate any custom
log record attributes we need with a default if the
record doesn't have the attribute.

This lets us use Formatters and formatter strings that
reference custom log record attributes without worrying about
the formatter throwing a an error when it tries to create
the log message.

ie, we can add a '%(remote_addr)s' record to the format string
without the formatter crashing because a regular logger doesnt
create that attribute on log records.

Currently it is attached to our default AnsibleLogger so that
records get the attribute when created.

But it could also be added to our log handlers to filter the
messages 'received' from loggers. If other loggers (without
our filters) are going to send messages to our log handlers,
DefaultAttributesFilter could be used there.

Or it could be added to both...

Add a process context log filter

This is a logging filter that adds info about the current
process that is missing from default log records.

cmd_name: basename of first command in sys.argv (because %(processName)s
          is often 'MainProcess' or similar)
cmd_line: The full command line (argv). Note that at the moment this is
          censored in anyway, so --password=hunter42 will show up in the
          logs if formatted to show 'cmd_line'
user: The username the process is running as.

Also added a LOG_INDEXER_FRIENDLT_FORMAT that uses key=value format
(logstash/splunk/etc generally can auto-parse those)

log the exit code of cli

split logger format strings into a module

Undecided if I want to keep logger/ this deep. May
change back to something flatter with less namespaces.

add support for a yaml logging config file

add queue_handler.py

add color_debug.py to formatters

color_debug was formerly known as debug_logger and live
in a different repo.

It is a logging.Formatter used in combo with a stdout/stderr
streamHandler to a console. It can colorize log messages based
on the log record attributes.

Most notably, it can set a color based on the thread, threadName,
process, or processName that tries to be different for each thread.
So all the log records from MainThread might be blue, while
QueueConsumerThread might be red.

The per thread/pid/etc color can be used with or without including
the info in the log message, but it is more useful to include it.

Add a Summary() class to base.py

Summary() is more or less a repr class for base.Base
and it's subclasses intended for use with logging.

It attempts to show the 'useful' info about a playbook
data type (it's type, it's name, and any non-default attributes)

Some event/context oriented logging for playbook/

This is a start of generating log events for things like
starting and ending playbooks/plays/blocks/tasks/handlers

long term it may be more useful to log these via a play context
aware logging adapter, to allow structed logging of play details.