Make sure file content is bytes
This is a workaround for an issue that only seems
to occur in unittests when using mock_open().
The data gets read always ends up as a str() type
on PY2/PY3 where it should be str/bytes respectively.

This causes dataloader._get_file_contents to fail when
it calls self._vault.is_encrypted() because it requires
a py2 str or PY3 bytes.