Add data item to AnsibleError
AnsibleError.data is a dict that can be used
to hold additional error context. The data there
is transfered to TaskResult objects when there
is an exception. Then callbacks can use the
extra data when displaying errors.

So far, that is just a field for captured
stderr from the ssh connection plugin, used
by the default callback plugin to display any
errors.

Note that items in the AnsibleError.data variable
need to be serializable (JSON, in particular) so
that resulting TaskResult items are serializable
and can be sent across multiprocessing queues.

Remove AnsibleSshConnectionFailure and use
AnsibleConnectionFailure with extra data instead.