Mv raising of connect exc to ssh.py:_run_once()
It was previously in _run() (and before that,
exec_command()). But _run_once is only called
one place, so let it handle raising the exception.

That simplifies the exception handling some, so
the code in _run() loses a couple of indention
levels.

The code for tracking attempts and pause/backoff
was extracted to a _retries_with_backoff() generator
method.

Also show the number of retries that will be attempted
in the output, including the final attempt.