Add exception key to unreachable task result
Update default.py callback to treat exception keys in
v2_runner_on_unreachable the same as v2_runner_on_fail
(ie, if < 'vvv', indicate there was a traceback, and on
'vvv'+, show the traceback).

Note the traceback in this case is a bit misleading, since
it was from an expected and handled exception.