Add DefaultAttributesFilter
This is a logging filter that will populate any custom
log record attributes we need with a default if the
record doesn't have the attribute.

This lets us use Formatters and formatter strings that
reference custom log record attributes without worrying about
the formatter throwing a an error when it tries to create
the log message.

ie, we can add a '%(remote_addr)s' record to the format string
without the formatter crashing because a regular logger doesnt
create that attribute on log records.

Currently it is attached to our default AnsibleLogger so that
records get the attribute when created.

But it could also be added to our log handlers to filter the
messages 'received' from loggers. If other loggers (without
our filters) are going to send messages to our log handlers,
DefaultAttributesFilter could be used there.

Or it could be added to both...