Add 'deprecated' module for tracking deprecations
Setup a module with a deprecation registry and
helper methods for checking deprecations.

Add a Deprecation class, and subclass it for
each deprecation type, including a label,
the version, removal status, a message, and
optionally, a 'mitigated' method to check
if any thing mitigates the deprecation warning
(for ex, enabling a config option to enable
the new behavior).

Tracks that warnings are ownly shown once.

Has hook for display callback to decouple it
from 'display' module.