add color_debug.py to formatters
color_debug was formerly known as debug_logger and live
in a different repo.

It is a logging.Formatter used in combo with a stdout/stderr
streamHandler to a console. It can colorize log messages based
on the log record attributes.

Most notably, it can set a color based on the thread, threadName,
process, or processName that tries to be different for each thread.
So all the log records from MainThread might be blue, while
QueueConsumerThread might be red.

The per thread/pid/etc color can be used with or without including
the info in the log message, but it is more useful to include it.