Allow a task to return an 'exception' field
Previously, the only way to get a task to show an
exception was if the tasks stderr started with the
string 'Traceback'

If a task result includes 'exception', it supercedes
the stderr output. module_stderr will still be populated.

Also, module_stdout, module_stderr, and rc would only
be populated with the task stdout was invalid JSON.
So now those are populated always, and it's up to the
callback to ignore them. That should make callbacks
more useful.