attempt to defer needing to import module packages after playbook parse
there is a bit of catch-22 with task load/parse. It happens
early, but it also wants to be able to verify something is a
a module name so it tries to import module package paths.
Since the rule for a valid task arg is basically
'an attribute or a module name'.

If we want to parse the playbook first, before setting up
any module paths, this is a problem. Why would you do that?
So that you could let a role provide its library. So you
could let a role provide the _entire_ library. Why would
you do that? so you can manage it by version/env/etc.

ie, if we load a role before attempting to load module paths,
the role could provide it. Potentially, the role spec/args
could alter what library gets setup. that depends on the
role being able to set its own library path though (unsure
if that currently works)