Use shorter names/fields for processName/threadName
It is very useful to have processName in the default debug
format. It is not as useful for threadName or thread (tid).
But most are 'MainProcess' or 'WorkerProcess-3' or 'MainThread'
which makes for long log lines.

So shorted the fields in the default logging format. And use shorter
names where we can choose the name. And abbrevieate 'MainProcess' ->
'MP' and 'MainThread' -> 'MT' via the process_context filter.

Now, everything in the default output before the logger name should
be a fixed width, and ~10 chars shorter than before. This makes the
default output a little easier to read.