use basic.bytes_to_human i/s self.module.pretty_bytes
AnsibleModule.pretty_bytes() is deprecated, and also just
a wrapper for a module_utils.basic.bytes_to_human.

This eliminates one requirement for facts classes to have
an AnsibleModule() instance. get_bin_path() and run_command()
are the remaining ones.

Goal is to decouple facts classes from an AnsibleModule() instance
to make them more generally useful and easier to test.