add a subclass for AnsibleFactCollector
subclass knows how to build its collector list, and
it based on a NestedFactsCollector parent class, so it
populates to 'ansible_facts' subspace.

Maybe that last bit is overkill. But already need
a collector that does the nesting and one that knows how
to munge dict keys (for Ohai/Facter) so maybe not.