extra get_pkg_mgr_facts -> system/pkg_mgr.py
Also, move the PKG_MGR map from Facts() to pkg_mgr.py
and also add a variant of it (just the pkg mgr names) to
constants.py so that action_write_locks.py can use it without
having to import from module_utils. Previously it was sort
of weirdly coupled.