Replace WrapperCollectors for Network/Hardware
add Collector to their base module.

And add some kluges because the impls of the Network/Virtual/Hardware
classes still expect self.facts and they mutate it. Need to
update to get them returning just a dict to update, then we can
remove the temp filtering of what collected_facts are passed in
(and/or returned)