WIP, fiddling with using fact_id and a platform id for matching collectors
really needs matcher objects instead of various muning of
tuples as dict keys

only add collectors that match the fact_ids and the platform_info
to the list of collectors used.

atm, a collectors platform_id will default to 'Generic', and
any platform matches 'Generic'

goal is to select collector classes including matching the
systems platform in collector.py, instead of relying on
metaclasses in hardware/*. To finish this, the various
Facts() subclasses will need to be replaced entirely with
Collector() subclasses.