gut most of Facts() subclasses for linux
and update the base versions to skip the imports
on get_all_subclasses etc.

Note: at the moment, not linux fact collection will
likely throw an error. The various platform specific
modules need a PlatformCollector() and adding it
to default_collectors

Once that is done, the facts/hardware/__init__.py etc
can be emptied