let plugins be dir based py module (foo/__init__.py)
let plugins be loaded from PLUGIN_PATH/foo/__init__.py

Stop filtering found plugins that are __init__.py. Since
that means PLUGIN_PATH/__init__.py will be loaded, the
loading would previously throw a
"Skipping plugin (%s) as it seems to be invalid" warning.

So add a mechanism for plugin_loader to detect when it is
loading a 'base' plugin module (filter/__init__.py for ex).
Do that by adding a '_base_module' attribute to the python module
containing the base plugin module and checking for it before
warning about invalid plugins. This is more or less equiv
to creating a abstract base class, but for our py modules.

(A cleaner approach would be to seperate the base plugin
modules and classes into a different dir containing just
implementations so the plugin loader would not stumble into
the base module/classes, but shrug...)