use --vault-id to indicate id and path/type
--vault-id=prompt  # prompt for default vault id password
--vault-id=myorg@prompt  # prompt for a vault_id named 'myorg'
--vault-id=a_password_file  # load ./a_password_file for default id
--vault-id=myorg@a_password_file # load file for 'myorg' vault id

create vault_ids from --vault-password-file args

This makes any --vault-password-file lower precendence than
a --vault-id, so will be the last secrets to be tried in that case

use --ask-vault-pass to add a default prompt vault_id

use 1.1 format for 'default' vault_id

Only use a vault_id if we specify one. So if none
is specified and C.DEFAULT_VAULT_IDENTITY is 'default'
we use the old format.

If we set a new DEFAULT_VAULT_IDENTITY, then the default will
use version 1.2

check for ';' in vault_ids

';' is the vault header seperator, so we just dont
allow it in vault_ids

intg tests for --vault-id

add a 1.2 format example to intg tests

add vault_secrets matchers