fix up base.yml enough to run doc generation
use name as the key here

the rst targets generated from the name aren't really
valid at the moment

An explicit label or id would be useful here. Otherwise
we will have to invent some canonical way to sluggify names into
valid identifiers if we want to :ref: them easily.

Replace the 'name' with the dict key string that is
the primary key here. That key/id is usually of the
form 'DEFAULT_SOMETHING_WHATEVER' though, which means
exposing an id that is otherwise an implementation detail.

TODO: assuming there are no name conflicts, try using the
ini section.key as the label. Or maybe even just ini key.

workaround missing 'name' fields