start adding a yaml !vault-plaintext object
To be used when vault editing yaml files with
embedded vaulted variables.

The !vault will be decrypted and replaced with a
!vault-plaintext object to be edited.

atm, the !vault-plaintext  is a map with the plaintext,
the vault-id to encrypt with, and the 'decrypted_from'
original vault content.

This is mostly so that vault-edit has enough context to
properly replace the original !vault with the new content
and reencrypt it with the right vault id (a single yaml
file could have vaulted vars that were encrypted with
different vault ids)