wip, FIXME: stub out a version of encrypt_value
TODO: figure out if we need to support this and do it, or
      just remove the support for $encrypted

Since we dont have a awx compat impl of Fernat256 to
apply here, stub out encrypt_value for now.

Not sure it makes sense to encrypt as a side effect of
validating, but either way it means this implementation is
not compatible with the awx equilivent.