task_executor was trying to find all action plugins
Previously, if task_executor.TaskExecutor._get_action_handle tried
to find an action plugin that has not been already loaded (or doesnt
exist),  it would try to find a action plugin with a name based on
a 'module_prefix' created by splitting the requested name on '_'.
For collections, this results in it looking for something
like 'somenamespace.some' if trying to find a
'somenamespace.some_name.some_action_name'.

This is incorrect except for the special case of some networking
action plugins, but the all() always runs all the expressions in
it, so a 'module_prefix in self._shared_loader_obj.action_loader'.

The 'module_prefix in self._shared_loader_obj.action_loader' results
on the action_loader running it's find_plugins, which potentially
causes the loader to raise inport/loader exceptions since a
module_prefix like 'somenamespace.some' is bogus and doesnt exist
and can't be loaded.

Replacing the all() with an 'and' expression that will short circuit
if not the network group module case fixes it. This prevents the
loader from trying to find a 'module_prefix' that will not exist.